Changed some things
Made working with capital turn it illiquid analogous to fatigue

Did not use the fatiguing aspect because it is used in the exile with cash going to husk

made exploring with lensing burn it

updated decks so they have consistent odds with streets. removed the restlessness from cemetary, odds explained here

2/7 chance of influence

2/7 chance of advancement (influence decays to advancement)

1/7 chance of negative influence

2/7 chance of positive influence

temple is doubled up since it has a higher variety under this. Lensing is considered a positive influence despite its danger

Brought investment job into here

added an icon for the temple